created selectAllPlayer method
