+++ src/main/java/net/scit/bluemarble/dao/bluemarbleDAO.java
@@ -1,5 +1,7 @@
 package net.scit.bluemarble.dao;
 
+import java.util.ArrayList;
+
 import org.apache.ibatis.annotations.Mapper;
 
 import net.scit.bluemarble.vo.Player;
@@ -7,4 +9,6 @@ import net.scit.bluemarble.vo.Player;
 @Mapper
 public interface bluemarbleDAO {
 	public Player selectOnePlayer(int playerid);
+
+	public ArrayList<Player> selectAllPlayer();
 }
