+++ src/main/java/net/scit/bluemarble/service/bluemarbleServiceImpl.java
@@ -1,5 +1,7 @@
 package net.scit.bluemarble.service;
 
+import java.util.ArrayList;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,6 +17,12 @@ public class bluemarbleServiceImpl implements bluemarbleService {
 	public Player selectOnePlayer(int playerid) {
 		Player player = dao.selectOnePlayer(playerid);
 		return player;
+	}
+
+	@Override
+	public ArrayList<Player> selectAllPlayer() {
+		ArrayList<Player> list = dao.selectAllPlayer();
+		return list;
 	}
 
 }
