+++ src/main/resources/mappers/bluemarble.xml
@@ -4,7 +4,13 @@
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="net.scit.bluemarble.dao.bluemarbleDAO">
+	<!-- 한 플레이어의 정보 조회 -->
 	<select id="selectOnePlayer" parameterType="int" resultType="Player">
-	SELECT * FROM BM_PLAYER_STATUS WHERE PLAYERID = #{playerid}
+		SELECT * FROM BM_PLAYER_STATUS WHERE PLAYERID = #{playerid}
+	</select>
+	
+	<!-- 모든 플레이어의 정보 조회 -->
+	<select id="selectAllPlayer" resultType="Player">
+		SELECT * FROM BM_PLAYER_STATUS
 	</select>
 </mapper>
